% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/augment.R
\name{augment.default}
\alias{augment.default}
\title{Default augment method}
\usage{
\method{augment}{default}(object, ...)
}
\arguments{
\item{object}{An object of unsupported class}

\item{...}{Ignored}
}
\value{
This method always throws an error
}
\description{
Default method that throws an error for unsupported object types.
Augmentation is only implemented for subspace learner objects.
}
