% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_LearnerSubspaceBoxGeneral.R
\name{autoplot.LearnerSubspaceBoxGeneral}
\alias{autoplot.LearnerSubspaceBoxGeneral}
\title{autoplot method for LearnerSubspaceBoxGeneral}
\usage{
\method{autoplot}{LearnerSubspaceBoxGeneral}(
  object,
  select = "all",
  force = FALSE,
  wrap = TRUE,
  size_top = 0.7,
  size_all = 0.5,
  ...
)
}
\arguments{
\item{object}{A LearnerSubspaceBoxGeneral object}

\item{select}{selected columns to print pairwise}

\item{force}{flag: Force wrapping?}

\item{wrap}{flag: Wrap plots using patchwork?}

\item{size_top}{point size of top_config points}

\item{size_all}{point size of all points in the dataset}

\item{...}{Additional arguments to patchwork::wrap_plots()}
}
\description{
autoplot method for LearnerSubspaceBoxGeneral
}
