% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_LearnerSubspaceBoxGeneral.R
\name{coef.LearnerSubspaceBoxGeneral}
\alias{coef.LearnerSubspaceBoxGeneral}
\title{coef method for LearnerSubspaceBoxGeneral}
\usage{
\method{coef}{LearnerSubspaceBoxGeneral}(object, ...)
}
\arguments{
\item{object}{A LearnerSubspaceBoxGeneral object}

\item{...}{Additional arguments (unused)}
}
\description{
coef method for LearnerSubspaceBoxGeneral
}
