% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_LearnerSubspaceElips.R
\name{coef.LearnerSubspaceElips}
\alias{coef.LearnerSubspaceElips}
\title{coef method for LearnerSubspaceElips}
\usage{
\method{coef}{LearnerSubspaceElips}(object, ...)
}
\arguments{
\item{object}{A LearnerSubspaceElips object}

\item{...}{Additional arguments (unused)}
}
\description{
coef method for LearnerSubspaceElips
}
