% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSubspaceBoxGeneral.R
\name{LearnerSubspaceBoxGeneral}
\alias{LearnerSubspaceBoxGeneral}
\title{LearnerSubspaceOrientedBox class}
\description{
LearnerSubspaceOrientedBox class

LearnerSubspaceOrientedBox class
}
\section{Super class}{
\code{\link[spacefinder:LearnerSubspace]{spacefinder::LearnerSubspace}} -> \code{LearnerSubspaceBoxGeneral}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSubspaceBoxGeneral-clone}{\code{LearnerSubspaceBoxGeneral$clone()}}
}
}
\if{html}{\out{
<details open><summary>Inherited methods</summary>
<ul>
<li><span class="pkg-link" data-pkg="spacefinder" data-topic="LearnerSubspace" data-id="initialize"><a href='../../spacefinder/html/LearnerSubspace.html#method-LearnerSubspace-initialize'><code>spacefinder::LearnerSubspace$initialize()</code></a></span></li>
<li><span class="pkg-link" data-pkg="spacefinder" data-topic="LearnerSubspace" data-id="train"><a href='../../spacefinder/html/LearnerSubspace.html#method-LearnerSubspace-train'><code>spacefinder::LearnerSubspace$train()</code></a></span></li>
</ul>
</details>
}}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSubspaceBoxGeneral-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSubspaceBoxGeneral-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSubspaceBoxGeneral$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
