% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_task_subspace.R
\name{as_task_subspace}
\alias{as_task_subspace}
\title{sugar function to create new TaskSubspace}
\usage{
as_task_subspace(...)
}
\arguments{
\item{...}{arguments passed to TaskSubsapce$new()}
}
\description{
sugar function to create new TaskSubspace
}
