% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/autoplot_LearnerSubspaceElips.R
\name{autoplot.LearnerSubspaceElips}
\alias{autoplot.LearnerSubspaceElips}
\title{autoplot method for LearnerSubspaceElips}
\usage{
\method{autoplot}{LearnerSubspaceElips}(object, size_top = 0.7, size_all = 0.5, wrap = TRUE, ...)
}
\arguments{
\item{object}{A LearnerSubspaceElips object}

\item{size_top}{point size of top_config points}

\item{size_all}{point size of all points in the dataset}

\item{wrap}{flag: Wrap plots using patchwork?}

\item{...}{Additional arguments (unused)}
}
\description{
autoplot method for LearnerSubspaceElips
}
