% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coef_LearnerSubspaceBox.R
\name{coef.LearnerSubspaceBox}
\alias{coef.LearnerSubspaceBox}
\title{coef method for LearnerSubspaceBox}
\usage{
\method{coef}{LearnerSubspaceBox}(object, ...)
}
\arguments{
\item{object}{A LearnerSubspaceBox object}

\item{...}{Additional arguments (unused)}
}
\description{
coef method for LearnerSubspaceBox
}
