% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LearnerSubspace.R
\name{LearnerSubspace}
\alias{LearnerSubspace}
\title{LearnerSubspace base class}
\description{
LearnerSubspace base class

LearnerSubspace base class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{task}}{A TaskSubspace object}

\item{\code{result}}{training result}

\item{\code{top_configs}}{the best hyperparameter configurations w.r.t. to q-value and target_measure}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LearnerSubspace-new}{\code{LearnerSubspace$new()}}
\item \href{#method-LearnerSubspace-train}{\code{LearnerSubspace$train()}}
\item \href{#method-LearnerSubspace-clone}{\code{LearnerSubspace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSubspace-new"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSubspace-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSubspace$new(task)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{task}}{A \code{TaskSubspace} object}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSubspace-train"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSubspace-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSubspace$train(
  q_val = 1,
  lambda = NULL,
  tasks = NULL,
  exclude_tasks = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{q_val}}{Quantile threshold for filtering configurations (0-1)}

\item{\code{lambda}}{Regularization parameter for slack variables (default = NULL, no slack)
Higher values allow more points outside the subspace
Agnostically compute subspaces}

\item{\code{tasks}}{Character vector of task names to include (optional, yet mutually exclusive with exclude_tasks)}

\item{\code{exclude_tasks}}{Character vector of task names to exclude (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LearnerSubspace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-LearnerSubspace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LearnerSubspace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
