% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outliers_LearnerSubspace.R
\name{outliers}
\alias{outliers}
\title{outliers method for learners inheriting from LearnerSubspace}
\usage{
outliers(object, ...)
}
\arguments{
\item{object}{An object inheriting from LearnerSubspace}

\item{...}{Additional arguments (unused)}
}
\description{
outliers method for learners inheriting from LearnerSubspace
}
