% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacefinder-package.R
\docType{package}
\name{spacefinder-package}
\alias{spacefinder}
\alias{spacefinder-package}
\title{spacefinder: Subspace Learning for Hyperparameter Optimization}
\description{
Learn promising hyperparameter subspaces from benchmark data using
geometric methods. The package provides learners for fitting axis-aligned
hyperrectangles, oriented hyperrectangles, and ellipsoids to high-performing
hyperparameter configurations.
}
\details{
\strong{Main Components:}
\itemize{
\item \code{\link{TaskSubspace}}: Define hyperparameter optimization tasks
\item \code{\link{LearnerSubspaceBox}}: Fit axis-aligned hyperrectangles
\item \code{\link{LearnerSubspacePolygon}}: Fit oriented hyperrectangles
\item \code{\link{LearnerSubspaceEllipsoid}}: Fit ellipsoids
}

\strong{Key Methods:}
\itemize{
\item \code{train()}: Fit subspace to top-performing configurations
\item \code{coef()}: Extract fitted subspace parameters
\item \code{augment()}: Add beta density parameters
\item \code{autoplot()}: Visualize fitted subspaces
\item \code{outliers()}: Extract outlier configurations
}
}
\seealso{
Useful links:
\itemize{
  \item \url{https://nikogerman.github.io/spacefinder/}
}

}
\author{
\strong{Maintainer}: Nikolai German \email{niko.german@gmail.com} (\href{https://orcid.org/0009-0001-7394-8367}{ORCID})

}
\keyword{internal}
