% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spacefinder-package.R
\docType{package}
\name{spacefinder-package}
\alias{spacefinder}
\alias{spacefinder-package}
\title{spacefinder: Find Promising Subspaces Within The Hyperparameter Space}
\description{
Uses historic multi-instance data to identify hyper-rectangles or hyper-ellipsoids in the hyperparameter space.
}
\author{
\strong{Maintainer}: Nikolai German \email{niko.german@gmail.com} (\href{https://orcid.org/0009-0001-7394-8367}{ORCID})

}
\keyword{internal}
