% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_LearnerSubspace.R
\name{summary.LearnerSubspace}
\alias{summary.LearnerSubspace}
\title{summary method for LearnerSubspace}
\usage{
\method{summary}{LearnerSubspace}(object, ...)
}
\arguments{
\item{object}{a object which inherits from LearnerSubspace}

\item{...}{Additional arguments (unused)}
}
\description{
summary method for LearnerSubspace
}
