% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TaskSubspace.R
\name{TaskSubspace}
\alias{TaskSubspace}
\title{TaskSubspace class}
\description{
TaskSubspace class

TaskSubspace class
}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{data}}{A data table, containing hyperparameter and a target measure.
needs to have a column named 'task'.}

\item{\code{target_measure}}{a character string: the target metric (e.g. auc, mce, brier loss,...)}

\item{\code{hps}}{One or more numeric hyperparameters}

\item{\code{cat_hps}}{currently up to one categorical hyperparamter}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-TaskSubspace-new}{\code{TaskSubspace$new()}}
\item \href{#method-TaskSubspace-clone}{\code{TaskSubspace$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskSubspace-new"></a>}}
\if{latex}{\out{\hypertarget{method-TaskSubspace-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskSubspace$new(
  data,
  formula = NULL,
  target_measure = NULL,
  hps = NULL,
  cat_hps = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{data}}{A data.table containing task performance data}

\item{\code{formula}}{target_measure ~ (hp1 + hp2 + ...) * (cat_hp1 + cat_hp2 + ...) (optional)}

\item{\code{target_measure}}{Name of the performance measure column}

\item{\code{hps}}{Character vector of continuous hyperparameter names}

\item{\code{cat_hps}}{Character vector of categorical hyperparameter names (optional)}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-TaskSubspace-clone"></a>}}
\if{latex}{\out{\hypertarget{method-TaskSubspace-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{TaskSubspace$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
